item, itemDropper
드롭템과 드롭퍼 만드는 중

+++ Assets/02.Scripts/SO/SkillDataSO.cs
@@ -12,5 +12,5 @@ public class SkillDataSO : ScriptableObject
     [ResizableTextArea]
     public string description;
 
-    public int coolTime;
+    public float coolTime;
 }
